PB-239: Remove unused function, apply DRY  on test case.
