+++ tests/cypress/tests-e2e/reportProblem.cy.js
@@ -2,51 +2,13 @@
 
 import { APP_VERSION } from '@/config'
 
+import { interceptFeedback, parseFormData } from './feedbackTestUtils'
+
 const text =
     'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.'
 const validEmail = 'this.is.a.valid@email.com'
 
 describe('Testing the report problem form', () => {
-    function interceptFeedback(success) {
-        cy.intercept('POST', '**/api/feedback', (req) => {
-            req.reply({
-                body: {
-                    success,
-                },
-                delay: 1000,
-            })
-        }).as('feedback')
-    }
-
-    function getFormDataBoundary(request) {
-        const contentType = request.headers['content-type']
-        const boundaryMatch = contentType.match(/boundary=([\w-]+)/)
-        return boundaryMatch && boundaryMatch[1]
-    }
-
-    function parseFormData(request) {
-        const boundary = getFormDataBoundary(request)
-        const formDataParts = request.body.split(boundary)
-        return (
-            formDataParts
-                .filter((part) => part.indexOf('Content-Disposition') !== -1)
-                .map((rawPart) => {
-                    const split = rawPart
-                        .split(/\r?\n/)
-                        .filter((split) => split !== '--' && split.length > 0)
-                    const name = split[0].substring(
-                        split[0].indexOf('name="') + 6, // removing the 6 chars from name="
-                        split[0].length - 1 // removing trailing "
-                    )
-                    return {
-                        [name]: split[1],
-                    }
-                })
-                // flatten array of params into a single object (with param name as keys)
-                .reduce((accumulator, param) => Object.assign(accumulator, param))
-        )
-    }
-
     function closeForm() {
         cy.get('[data-cy="modal-close-button"]').should('be.visible').click()
     }
@@ -105,7 +67,7 @@ describe('Testing the report problem form', () => {
         cy.log('It generates a complete request to service-feedback')
         openForm()
         interceptFeedback(true)
-        // cy.get('[data-cy="feedback-email"]').type(validEmail)
+        cy.get('[data-cy="feedback-email"]').type(validEmail)
         cy.get('[data-cy="report-problem-text"]').type(text)
 
         cy.get('[data-cy="submit-report-problem-button"]').click()
@@ -118,7 +80,7 @@ describe('Testing the report problem form', () => {
                 { name: 'feedback', contains: text },
                 { name: 'version', contains: APP_VERSION },
                 { name: 'ua', contains: navigator.userAgent },
-                // { name: 'email', contains: validEmail },
+                { name: 'email', contains: validEmail },
             ].forEach((param) => {
                 expect(interception.request.body).to.be.a('String')
                 const formData = parseFormData(interception.request)
